The Progression Cloak. (#451) (#295)
<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

# Description

<!--
Explain this PR in as much detail as applicable

Some example prompts to consider:
How might this affect the game? The codebase?
What might be some alternatives to this?
How/Who does this benefit/hurt [the game/codebase]?
-->

Cherry-pick of the Progress Cape from Funky-Station

https://github.com/funky-station/funky-station/pull/451

---

<!--
This is default collapsed, readers click to expand it and see all your
media
The PR media section can get very large at times, so this is a good way
to keep it clean
The title is written using HTML tags
The title must be within the <summary> tags or you won't see it
-->

<details><summary><h1>Media</h1></summary>
<p>

![image](https://github.com/user-attachments/assets/2730c129-80c8-42a9-b62c-94d39136b1dc)
Image provided by original PR

</p>
</details>

---

# Changelog

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl: Icepicked
- add: Added Progress Cape from Funky-Station to Pride-O-Mat

Co-authored-by: Icepick <122653407+Icepicked@users.noreply.github.com>